+++ pages/role/named-person-mvp/mandatory-training/content/index.js
