+++ pages/role/named-person-mvp/mandatory-training/index.js
