+++ pages/role/named-person-mvp/mandatory-training/views/index.jsx
